requirements.txt
# RAG-Powered Compliance Advisor for African SMEs

# Streamlit UI
streamlit>=1.32.0

# LangChain ecosystem
langchain>=1.2.0
langchain-core>=1.2.0
langchain-text-splitters>=0.2.0
langchain-openai

# LangGraph
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0

# Vector database
chromadb>=0.4.0

# Retrieval & reranking
sentence-transformers>=2.2.0
rank_bm25>=0.2.0

# Web scraping (for live document fetching)
requests>=2.28.0
beautifulsoup4>=4.12.0

# Environment management
python-dotenv>=1.0.0

# Data processing
pydantic>=2.0.0
